Simplify PumiceDBClient.Write() and writeKV(), change method owner
The public Write() method is now a wrapper of writeKV().

writeKV() has undergone a lot of changes which intend to more clearly
delegate functionality, especially wrt to Encoding.

The methods are now owned by PmdbReqArgs instead of PmdbClientObj. The
PmdbClientObj pointer is now contained inside of PmdbReqArgs for the
few instances where it's needed.

PmdbReqArgs::encodeRequestPayload() conditionally encodes the user's
part of the buffer if encoded contents are not yet detected.

PmdbReqArgs::encodeRequest() encodes the entire request and must be
called after encodeRequestPayload().  This function returns the items
formerly returned by getPmdbReq() (which should be deprecated soon..).

PmdbReqArgs::getObjectID() does the job of filling a pointer of
C.pmdb_obj_id_t.

PmdbReqArgs::writeKV() calls the above private methods to prepare for
the issuance of C.PmdbObjPut().